Detached branch (#94)
* v1.2.0 for lighter install.

* update main-ci.yml

* fixed test_dataset
