+++ .github/workflows/main-ci.yml
@@ -29,7 +29,7 @@ jobs:
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        python -m pip install flake8 pytest setuptools wheel twine
+        python -m pip install flake8 pytest[torch] setuptools wheel twine
         pip install -e .
         pip install -r requirements.txt
     - name: Import checking
@@ -37,5 +37,5 @@ jobs:
         python -c "import imgtools"
     - name: Run pytest
       run: |
-        pytest tests -s
+        pytest tests
